ui: move min-zoom from a global topbar setting to a per-canvas overlay
Each session canvas now has its own min-zoom value, persisted by slot
(stable across reloads). A small "🔍" overlay button at the canvas's
top-right corner shows the current minimum (or "⊘" when off) and opens
a popover with a slider — same range as before (off / 0–200% step 5%,
plus a Reset button).

- new web/components/agent-visualizer/canvas-zoom-control.tsx wraps the
  button + portal popover; the portal escapes the FloatingPanel's
  overflow:hidden.
- session-canvas-panel.tsx pulls minZoomLevel from a per-slot, per-
  instance usePersistedState and passes it to AgentCanvas.
- index.tsx and top-bar.tsx no longer carry the global setting.
