+++ web/components/agent-visualizer/top-bar.tsx
@@ -138,11 +138,6 @@ export interface TopBarProps {
    *  chip per id so they can be reopened via onShowCanvas. */
   hiddenCanvases?: ReadonlySet<string>
   onShowCanvas?: (sessionId: string) => void
-  /** Floor for the canvas auto-fit scale, 0 = no minimum (current default).
-   *  Exposed in a small slider popover so the user can prevent the camera
-   *  from zooming out too far when many agents are spread across the canvas. */
-  minZoomLevel?: number
-  onMinZoomLevelChange?: (next: number) => void
   onTogglePanel: (panel: 'files' | 'transcript' | 'cost' | 'messages') => void
   onToggleTimeline: () => void
   onToggleMute: () => void
@@ -158,7 +153,6 @@ export const TopBar = memo(function TopBar({
   showFileAttention, showTranscript, showCostOverlay, showTimeline, showMessageFeed, isMuted,
   workspaceFilter,
   hiddenCanvases, onShowCanvas,
-  minZoomLevel = 0, onMinZoomLevelChange,
   onTogglePanel, onToggleTimeline, onToggleMute, onUiClick,
 }: TopBarProps) {
   const { resetLayout, saveLayout, hardResetLayout, tilePanels, instanceId, hostId, otherInstances } = usePanelLayout()
@@ -264,9 +258,6 @@ export const TopBar = memo(function TopBar({
             sessions={allSessions}
             onRemoveSession={onRemoveSession}
           />
-          {onMinZoomLevelChange && (
-            <MinZoomButton value={minZoomLevel} onChange={onMinZoomLevelChange} />
-          )}
           {hiddenCanvases && hiddenCanvases.size > 0 && onShowCanvas && (
             <ClosedCanvasChips
               hiddenIds={hiddenCanvases}
@@ -503,124 +494,6 @@ function WorkspaceFilterButton({
   )
 }
 
-// ─── Min-zoom button + popover ──────────────────────────────────────────────
-// A small slider that floors the auto-fit scale on every canvas. 0 = off
-// (current default — auto-fit picks any zoom that fits the bounding box).
-
-function MinZoomButton({ value, onChange }: { value: number; onChange: (v: number) => void }) {
-  const [open, setOpen] = useState(false)
-  const buttonRef = useRef<HTMLDivElement>(null)
-  const popoverRef = useRef<HTMLDivElement>(null)
-  const [anchorRect, setAnchorRect] = useState<{ top: number; right: number } | null>(null)
-
-  useEffect(() => {
-    if (!open) { setAnchorRect(null); return }
-    const update = () => {
-      const r = buttonRef.current?.getBoundingClientRect()
-      if (r) setAnchorRect({ top: r.bottom + 6, right: window.innerWidth - r.right })
-    }
-    update()
-    window.addEventListener('resize', update)
-    window.addEventListener('scroll', update, true)
-    return () => {
-      window.removeEventListener('resize', update)
-      window.removeEventListener('scroll', update, true)
-    }
-  }, [open])
-
-  useEffect(() => {
-    if (!open) return
-    const handler = (e: MouseEvent) => {
-      const target = e.target as Node
-      if (buttonRef.current?.contains(target)) return
-      if (popoverRef.current?.contains(target)) return
-      setOpen(false)
-    }
-    window.addEventListener('mousedown', handler)
-    return () => window.removeEventListener('mousedown', handler)
-  }, [open])
-
-  const pct = Math.round(value * 100)
-  const label = value > 0 ? `Min Zoom ${pct}%` : 'Min Zoom'
-  const active = value > 0
-
-  const popoverContent = (
-    <div
-      ref={popoverRef}
-      style={{
-        position: 'fixed',
-        top: anchorRect?.top ?? 0,
-        right: anchorRect?.right ?? 0,
-        zIndex: 10000,
-        minWidth: 240,
-        padding: 12,
-        background: COLORS.glassBg,
-        backdropFilter: 'blur(20px)',
-        WebkitBackdropFilter: 'blur(20px)',
-        border: `1px solid ${COLORS.glassBorder}`,
-        borderRadius: 8,
-        boxShadow: '0 8px 24px rgba(0,0,0,0.4)',
-      }}
-    >
-      <div style={{ fontSize: 11, color: COLORS.textPrimary, marginBottom: 8, display: 'flex', justifyContent: 'space-between' }}>
-        <span>Auto-fit floor</span>
-        <span style={{ color: COLORS.textMuted }}>{value > 0 ? `${pct}%` : 'off'}</span>
-      </div>
-      <input
-        type="range"
-        min={0}
-        max={200}
-        step={5}
-        value={pct}
-        onChange={(e) => onChange(Number(e.target.value) / 100)}
-        style={{ width: '100%', accentColor: COLORS.holoBase }}
-        aria-label="Minimum auto-fit zoom level"
-      />
-      <div style={{ display: 'flex', justifyContent: 'space-between', fontSize: 9, color: COLORS.textMuted, marginTop: 4 }}>
-        <span>off</span>
-        <span>50%</span>
-        <span>100%</span>
-        <span>150%</span>
-        <span>200%</span>
-      </div>
-      <div style={{ marginTop: 10, fontSize: 10, color: COLORS.textMuted, lineHeight: 1.4 }}>
-        Stops the canvas from zooming out below this level when fitting all
-        agents. Wheel zoom is unaffected. Set to 200% to effectively lock
-        zoom at the max.
-      </div>
-      <div style={{ display: 'flex', justifyContent: 'flex-end', marginTop: 10 }}>
-        <button
-          onClick={() => onChange(0)}
-          disabled={value === 0}
-          style={{
-            padding: '4px 10px',
-            fontSize: 10,
-            fontFamily: 'monospace',
-            color: value === 0 ? COLORS.textMuted : COLORS.textPrimary,
-            background: 'transparent',
-            border: `1px solid ${COLORS.toggleBorder}`,
-            borderRadius: 4,
-            cursor: value === 0 ? 'default' : 'pointer',
-            opacity: value === 0 ? 0.5 : 1,
-          }}
-        >
-          Reset
-        </button>
-      </div>
-    </div>
-  )
-
-  return (
-    <>
-      <div ref={buttonRef}>
-        <ToggleButton active={active} onClick={() => setOpen(o => !o)}>
-          {label}
-        </ToggleButton>
-      </div>
-      {open && anchorRect && typeof document !== 'undefined' && createPortal(popoverContent, document.body)}
-    </>
-  )
-}
 
 // ─── Closed-canvas chips ────────────────────────────────────────────────────
 // Renders one chip per ✕-closed canvas. Click to reopen. Sits next to the
